refactor: ioredis instrumentation refactored moved traceConnection to instrumentation.ts refactored tests to use setConfig instead of new instrumentation
